Chore: Cleaned up code from May Comp
