--- control/scripts/conf/max_wheel_speed_levels.json
@@ -1,25 +1,0 @@
-{
-    "max_speeds": [
-      {
-        "gear": 1,
-        "speed": 2.0
-      },
-      {
-        "gear": 2,
-        "speed": 4.0
-      },
-      {
-        "gear": 3,
-        "speed": 6.0
-      },
-      {
-        "gear": 4,
-        "speed": 8.0
-      },
-      {
-        "gear": 5,
-        "speed": 10.0
-      }
-    ]
-  }
-
