+++ control/scripts/pantilt_control_node.py
@@ -33,6 +33,8 @@ class pantilt(Node):
         # Get IMU data and publish it
         imu_data = self.pantilt_firmware.get_imu_data()
         imu_msg = Float32MultiArray()
+        for i in range(len(imu_data)):
+            imu_data[i] = float(imu_data[i])
         imu_msg.data = tuple(imu_data)
         self.imu_publisher.publish(imu_msg)
         # Get GPS data and publish it
