+++ control/setup.py
@@ -31,7 +31,6 @@ setup(
             "odrive_node = scripts.node_odrive_drive:main",
             "gui_node = scripts.node_control_gui:main",
             "drive_firmware_node = scripts.drive_firmware_node:main",
-            "driveCAN_firmware_node = scripts.driveCAN_firmware_node:main",
             "pantilt_control_node = scripts.pantilt_control_node:main"
         ],
     },
